chore: extract inner class
