+++ src/main/java/com/example/demo/Page.java
@@ -1,0 +1,4 @@
+package com.example.demo;
+
+record Page(String id, String title, String content, String parentId) {
+}
