+++ src/main/java/com/example/demo/PageViewRequest.java
@@ -1,0 +1,4 @@
+package com.example.demo;
+
+record PageViewRequest(String pageId) {
+}
