+++ src/main/java/com/example/demo/SummaryPageInfo.java
@@ -1,0 +1,4 @@
+package com.example.demo;
+
+record SummaryPageInfo(String id, String title, String parentId) {
+}
